+++ client/public/index.html
@@ -3,7 +3,6 @@
   <head>
     <meta charset="UTF-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1, shrink-to-fit=no" />
-    <script src="https://kit.fontawesome.com/09fc819f23.js" crossorigin="anonymous"></script>
     <title>Engage Starter</title>
   </head>
