+++ client/src/components/hooks/useSquidList.jsx
@@ -4,7 +4,7 @@ import { ApiClient } from "../../backend/ApiClient"
 
 export const useSquidList = (pageOffset, pageSize) =>
   useQuery(
-    ["squids", { pageOffset }],
+    ["squids", { pageOffset, pageSize }],
     () =>
       ApiClient.get("/squids", { params: { pageOffset, pageSize } }).then((res) => res.data.squids),
     { keepPreviousData: true }
