+++ client/src/style/index.pcss
@@ -1,31 +1,35 @@
-.squids-list__header {
-  @apply font-eczar text-center text-5xl font-semibold py-3;
-}
+.squids-list {
+  @apply mx-24;
+  
+  & .squids-list__header {
+    @apply font-eczar text-center text-5xl font-semibold py-3;
+  }
 
-.squids-list__tile-header {
-  @apply font-medium;
-}
+  & .squids-list__tile-header {
+    @apply font-medium;
+  }
 
-.squids-list__tile {
-  @apply my-3 text-lg;
+  & .squids-list__tile {
+    @apply my-3 text-lg;
+  }
 }
 
 .squids-pagination {
-  @apply text-center text-xl mt-10;
-}
+  @apply text-center text-xl mt-10 mb-5;
 
-.squids-pagination__scroll-icon {
-  @apply px-2 text-cyan-600 hover:text-cyan-500 transition duration-200;
-}
-
-.squids-pagination__scroll-icon__disabled {
-  @apply text-slate-300 pointer-events-none;
-}
-
-.squids-pagination__page-number {
-  @apply px-3 hover:text-cyan-500 transition duration-200;
-}
-
-.squids-pagination__current-page-number {
-  @apply font-medium px-3 shadow-lg rounded-md text-zinc-200 bg-cyan-700;
+  & .squids-pagination__scroll-icon {
+    @apply px-2 text-cyan-600 hover:text-cyan-500 transition duration-200;
+  }
+  
+  & .squids-pagination__scroll-icon__disabled {
+    @apply text-slate-300 pointer-events-none;
+  }
+  
+  & .squids-pagination__page-number {
+    @apply px-3 hover:text-cyan-500 transition duration-200;
+  }
+  
+  & .squids-pagination__current-page-number {
+    @apply font-medium px-3 shadow-lg rounded-md text-zinc-200 bg-cyan-700;
+  }
 }
