+++ client/src/style/main.pcss
@@ -13,10 +13,6 @@
   }
 }
 
-.squids-list {
-  @apply mx-24;
-}
-
 /* .something {
   @apply bg-red-500;
   & .something-else {
