+++ client/src/style/topBar.pcss
@@ -1,12 +1,13 @@
 .menu {
   @apply flex flex-row bg-cyan-700 text-zinc-200 py-2;
+  
   & li {
     @apply font-special ml-8 my-3 text-xl;
   }
-}
-
-.menu__nav-link {
-  @apply cursor-pointer hover:text-zinc-400 transition duration-200;
+  
+  & .menu__nav-link {
+    @apply cursor-pointer hover:text-zinc-400 transition duration-200;
+  }
 }
 
 .top-bar-left {
